Set action='append' on --file argument
Set action=‘append’ when adding the argument parser for --file (-f), so
that it can accumulate multiple files.

If the length of args.file is 0 (i.e., no file was specified), then
attempt to open docker-compose.yml as the file. docker-compose.yml is
no longer set as the argument's default value  because action=append
adds to it instead of overriding it.

For now, the first (0-index) file is used to determine compose_base_dir
and create the DockerCompose object.