Specifically check for service in docker service ls
The name filter on docker service ls matches on all or part of a task's
name. This means that is_service_exists will erroneously return True if
the service name is a substring of another existing service.

Use awk to extract just the names from docker service ls, then grep for
a line that specifically matches the service being checked for
existence.

Resolves #14.